projects/qr-code-encoder-decoder/requirements.txt
Pillow==11.3.0
pyzbar==0.1.9
qrcode==8.2
